fix: correcting function definitions

+++ src/get_stats/get_stats.h
@@ -4,7 +4,7 @@
 int get_gpu_freq();
 int get_gpu_temp();
 char* get_gpu_stats();
-int truncate_gpu_freq(int freq);
-int truncate_gpu_temp(int temp);
+int truncate_gpu_freq(double freq);
+int truncate_gpu_temp(double temp);
 
 #endif
